@eme64 2023-01-12
@vnkozlov Something I failed to mention in the description:
In [ByteBufferTest.java](https://github.com/openjdk/jdk/blob/master/test/hotspot/jtreg/compiler/intrinsics/unsafe/ByteBufferTest.java) (which is the implementation of both tests) we have lots of inner loops, which all have at least `100` iterations. The buffer has a length of `1024`, the element size is maximally `8` bytes. In those inner loops, we do all the buffer-operations, which lead to the `Unsafe` intrinsics.
So then if we take even the OSR C2 threshold of `Tier4BackEdgeThreshold=40000`, I would assume that an outer iteration count of `40_000 / 100 = 400` would have to suffice.
I analyzed the `PrintInlining` statements, and it seems to really be ok for `5000`, also for `2000` in most cases.