@eme64 2023-01-16
@vnkozlov I filed a follow up RFE with your suggestion for improvement: [JDK-8300187](https://bugs.openjdk.org/browse/JDK-8300187)

Thanks for the help and reviews @vnkozlov @TobiHartmann 

/integrate